+++ C++/basic maths/AsciiVal.cpp
