+++ C++/basic maths/Calculate Power of a Number
