+++ C++/basic maths/CheckPowerOfTwo.cpp
