+++ C++/basic maths/Factorial_sum.cpp
