+++ C++/basic maths/Fibonacci_with_recursion.cpp
@@ -1,0 +1,21 @@
+#include<iostream>    
+using namespace std;      
+void printFibonacci(int n){    
+    static int n1=0, n2=1, n3;    
+    if(n>0){    
+         n3 = n1 + n2;    
+         n1 = n2;    
+         n2 = n3;    
+ cout<<n3<<" ";    
+         printFibonacci(n-1);    
+    }    
+}    
+int main(){    
+    int n;    
+    cout<<"Enter the number of elements: ";    
+    cin>>n;    
+    cout<<"Fibonacci Series: ";    
+    cout<<"0 "<<"1 ";  
+    printFibonacci(n-2); 
+     return 0;  
+}
