+++ C++/basic maths/GCD.cpp
