+++ C++/basic maths/Power_of_2.cpp
