+++ C++/basic maths/bitwiseoperator.cpp
