+++ C++/basic maths/factroial.cpp
