+++ C++/basic maths/power.cpp
