--- C++/spell check
@@ -1,41 +1,0 @@
-# Codeforces solution of spell check
-
-
-#include <bits/stdc++.h>
-using namespace std;
- 
- 
-void ayush(string s1){
-    long long int n;
-    cin>>n;
-    string s;
-    cin>>s;
-    if(n!=s1.size()){
-        cout<<"NO\n";
-        return;
-    }
-    sort(s.begin(),s.end());
-    for(int i=0;i<n;i++){
-        if(s[i]!=s1[i]){
-            cout<<"NO\n";
-            return;
-        }
-    }
-    cout<<"YES\n";
-}
- 
- 
-int main() {
-	// your code goes here
-	ios_base::sync_with_stdio(false);
-    cin.tie(NULL);  
-    cout.tie(NULL);
-	long long int t;
-	cin>>t;
-	string s1 = "Timur";
-	sort(s1.begin(),s1.end());
-	while(t--){
-	    ayush(s1);
-	}
-	return 0;
-}
